+++ thoth-api/src/graphql/model.rs
@@ -2211,10 +2211,12 @@ impl Work {
         self.cover_caption.as_ref()
     }
 
+    #[graphql(description = "Date and time at which the work record was created")]
     pub fn created_at(&self) -> Timestamp {
         self.created_at.clone()
     }
 
+    #[graphql(description = "Date and time at which the work record was last updated")]
     pub fn updated_at(&self) -> Timestamp {
         self.updated_at.clone()
     }
@@ -2234,6 +2236,13 @@ impl Work {
     )]
     pub fn page_interval(&self) -> Option<&String> {
         self.page_interval.as_ref()
+    }
+
+    #[graphql(
+        description = "Date and time at which the work record or any of its linked records was last updated"
+    )]
+    pub fn updated_at_with_relations(&self) -> Timestamp {
+        self.updated_at_with_relations.clone()
     }
 
     pub fn imprint(&self, context: &Context) -> FieldResult<Imprint> {