+++ thoth-export-server/src/xml/onix3_oapen.rs
@@ -28,7 +28,7 @@ impl XmlSpecification for Onix3Oapen {
                         w.write(XmlEvent::Characters("Thoth")).map_err(|e| e.into())
                     })?;
                     write_element_block("EmailAddress", w, |w| {
-                        w.write(XmlEvent::Characters("info@thoth.pub"))
+                        w.write(XmlEvent::Characters("distribution@thoth.pub"))
                             .map_err(|e| e.into())
                     })
                 })?;